+++ ichnaea/content/models.py
@@ -20,13 +20,13 @@ SCORE_TYPE_INVERSE = dict((v, k) for k, v in SCORE_TYPE.items())
 
 STAT_TYPE = {
     '': -1,
-    # 0 was location
+    # 'location': 0,
     'cell': 1,
     'unique_cell': 2,
     'wifi': 3,
     'unique_wifi': 4,
-    'deleted_cell': 5,
-    'deleted_wifi': 6,
+    # 'deleted_cell': 5,
+    # 'deleted_wifi': 6,
 }
 STAT_TYPE_INVERSE = dict((v, k) for k, v in STAT_TYPE.items())